Finished this feature
4. Ascension Drop Change:
Drop rate will now increase based on how many of that specific monster have been farmed since the last time that that type of monster dropped an Ascension Orb. For example, a slime will have a higher chance to drop an Ascension Orb if many slimes have been slain since the last time a slime dropped an orb. Each type of monster is tracked separately and that monster's counter is reset when it drops an orb for anyone.
An optional kill counter has been added to the experience ui. Toggle with /expkillcount (defaults to OFF). Shows the counter for the most recently killed monster type only

+++ Abilities/Abilities.cs
@@ -97,7 +97,7 @@ namespace ExperienceAndClasses.Abilities
             }
             protected override RETURN UseEffects()
             {
-                float value = ExperienceAndClasses.localMyPlayer.effectiveLevel * ExperienceAndClasses.localMyPlayer.healRate;
+                int value = (int)(ExperienceAndClasses.localMyPlayer.effectiveLevel * ExperienceAndClasses.localMyPlayer.healRate);
                 Vector2 target = Main.MouseWorld;
                 Projectile.NewProjectile(Main.LocalPlayer.position, new Vector2(0f), ExperienceAndClasses.mod.ProjectileType< AbilityProj.HealProj.Initial>(), value, 0, Main.LocalPlayer.whoAmI, target.X, target.Y);
